+++ packages/analytics-js-cookies/.size-limit.js
@@ -1,0 +1,11 @@
+/**
+ * DO NOT EDIT the size limits configured in this file.
+ * Should you update them, please contact CODEOWNERS.
+ */
+module.exports = [
+  {
+    name: 'Cookies Utilities',
+    path: 'dist/**/*.js',
+    limit: '2 KiB',
+  },
+];