+++ packages/loading-scripts/CHANGELOG.md
@@ -2,6 +2,11 @@
 
 This file was generated using [@jscutlery/semver](https://github.com/jscutlery/semver).
 
+## [3.0.12](https://github.com/rudderlabs/rudder-sdk-js/compare/@rudderstack/analytics-js-loading-scripts@3.0.11...@rudderstack/analytics-js-loading-scripts@3.0.12) (2024-06-25)
+
+### Dependency Updates
+
+* `@rudderstack/analytics-js` updated to version `3.0.11`
 ## [3.0.11](https://github.com/rudderlabs/rudder-sdk-js/compare/@rudderstack/analytics-js-loading-scripts@3.0.10...@rudderstack/analytics-js-loading-scripts@3.0.11) (2024-06-21)
 
 ### Dependency Updates